requirements.txt
click==8.1.8
cmake==3.31.4
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
mediapipe==0.10.11
numpy==1.24.2
opencv-python==4.11.0.86
pillow==11.1.0
